raphael: overlay: Disable proximity usage during doze
When the proximity sensor is in use, the smp2p_sleepstate
IRQ is fired multiple times a second, with each one
holding a 200ms wakelock. This is probably a bug in the
DSP firmware. To fix this, avoid using the proximity sensor
in doze mode, because sleep is preferred to turning off the
screen.

Signed-off-by: SageOfD6Path <mail2anirban95@gmail.com>